Enum hls_core::network::Request

pub enum Request {
    AcquireVolume {
        volume: VolumeId,
    },
    RefreshVolume {
        key: VolumeKey,
        volume: VolumeId,
        final_count: u64,
    },
    FreeVolume {
        key: VolumeKey,
        volume: VolumeId,
        latest_state: VolumeMetadata,
    },
    ReadChunk {
        key: VolumeKey,
        volume: VolumeId,
        path: (),
        chunk: ChunkId,
    },
    WriteChunk {
        key: VolumeKey,
        volume: VolumeId,
        path: (),
        chunk: ChunkId,
        data: (),
        new: bool,
    },
    DeleteChunk {
        key: VolumeKey,
        volume: VolumeId,
        path: FilePath,
        chunk: ChunkId,
    },
    DeleteFile {
        key: VolumeKey,
        volume: VolumeId,
        path: FilePath,
    },
}

The Requests send from the Client to the Server

Variants

AcquireVolume

Fields

volume: VolumeId

The Volume that the client wants to acquire

A Client wants to acquire a Volume to access it

RefreshVolume

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The relevant Volume

final_count: u64

The final UpdateCount of the Volumes CRDT

Refresh the lease for the Volume

FreeVolume

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The Volume to free

latest_state: VolumeMetadata

The Latest Volume Metadata State from the Client Side

Free a Volume again

ReadChunk

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The Volume in which the File/Chunk is located

path: ()

The Path specifying the File

chunk: ChunkId

The Chunk that should be read from said file

Read the Chunk of the specified File

WriteChunk

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The Volume in which the File/Chunk is located

path: ()

The Path of the File

chunk: ChunkId

The Chunk of the File

data: ()

The Data to write to the chunk

new: bool

If this is a new Chunk that should be created on the Server Side

Write the Chunk of the specified File

DeleteChunk

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The Volume in which the File/Chunk is located

path: FilePath

The Path of the File

chunk: ChunkId

The Chunk to delete

Delete the given Chunk

DeleteFile

Fields

key: VolumeKey

The Key to authenticate

volume: VolumeId

The Volume in which the File is located

path: FilePath

The Path of the File to delete

Delete the entire File, including all the Chunks